Add `nip06` feature
