+++ Cargo.toml
@@ -14,7 +14,8 @@ path = "uniffi-bindgen.rs"
 required-features = ["uniffi-cli"]
 
 [features]
-default = ["lmdb"]
+default = ["nip06","lmdb"]
+nip06 = ["nostr/nip06"]
 tor = ["nostr-connect/tor", "nostr-sdk/tor", "nwc/tor"]
 lmdb = ["nostr-sdk/lmdb"]
 ndb = ["nostr-sdk/ndb"]
